Fix bug when reading system.parts using UUID (issue 61220).
During the predicates pushing down of system.parts table, there is a
logic to use that predicates to filter on the storages. This makes sense
if the predicate is on the `database`, `table`, `engine`. But it can be
problem if the predicate is about `uuid`, because storage UUID is
apparently different from parts UUID.

Rename the column name from `uuid` to `storage_uuid` fixed this.