Limit the array index of FixedHashTable by min/max
If the type of key is 8 bits or 16 bits in aggregation, ClickHouse will use array
of 256 or 65536 length to store the key and boost the mergeSingleLevel, rather than
key comparison. However, if the key has occupied only small range of the total 65536
cells, most of the cycles are wasted on the `isZero()` to find the next cell which
is not zero in iterator++.
The solution is to use min/max and update min/max when emplace. Then we can set the
upper searching limit to max in iterator++. And just set min as the value of `begin()`,
rather than searching the first cell that not equals to 0.
We have tested the patch on 2x80 vCPUs server, Query 7 of ClickBench has gained 2.1x
performance improvement.

Signed-off-by: Jiebin Sun <jiebin.sun@intel.com>