fix infinite loop in AzureObjectStorage::listObjects
While the exit condition is correct for `AzureIteratorAsync`,
listObject may never exit.
This is because BlobContainerClient::LstBlobs sets a page, thus
making a sequential HasPage() condition true.

This issue reproduces when passing an empty path.
Tested with the following integration test:
https://pastila.nl/?01691b54/9ceaf103c91ab9e38ea3541abfcd1ae6#hUtrrz5sFHcBSHCLS9AlVA==
(will be added when plain_rewritable is enabled for all disk types).