Add endpoint_subpath S3 setting
Multiple plain_rewritable disks can share the same
endpoint, e.g., in the cloud system tables configuration.

At startup, a disk traverses the contents of a bucket to construct the
local-to-remote plain_rewritable mapping. A data race occurs when
another disk writes to the same bucket while the first disk is starting
up.

Introduce `endpoint_subpath` S3 setting, that will be unique for each
Pod instance in the cloud scenario.