Fix type inference for float (in case of small buffer)
In case of small buffer (i.e. --max_read_buffer_size 1) the pos() will
be always point to this one byte, so, comparing pos() will be always
evaluated to true.

And we cannot use count() as well, since in case of big buffer it will
be the same, plus, in case of reading extra byte for checking for '.'
the count() will be different, but it does not mean that the byte had
been interpreted (and allowing 1 byte of difference will not work almost
always, since it will read max_read_buffer_size bytes).

So instead, expose the has_fractional flag from the read helpers for
float, via two new methods:
- tryReadFloatTextExt
- tryReadFloatTextExtNoExponent

Where "ext" stands for "extended", which means expose extra information.

v2: consider number as float if it has '.' or 'e' (previously only if it
has some signs after those two it had been considered as float)

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>