Ignore async_load_databases for ATTACH query
It is quite odd that when ATTACH finishes the tables may not be exists,
due to async_load_databases.

For server startup it makes total sense, but not for queries.

Plus, you can execute queries in parallel if you want to make it faster.

Note, that server startup does not uses this code, see loadMetadata.cpp.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>