Remove the 47ms interval query because it's flaky
Even though running 5000 times the test in parallel
with 32 jobs in my dev machine doesn't show any error,
the CI does complain about the stddev of this one.
So, rather than increasing the range for all values
let's remove this one because for time-critical things
depending on the resources the result may be flaky.