Simplify logic using BackgroundSchedulePool
- Use the common schedule pool to collect the data, instead
of having a somewhat duplicate logic to do the same in a
single thread. This is better for code maintenance and also
for performance, since now there's a myriad of threads where
any of them may do the collection. Of course, I did not know
of the existence of the schedule pool before :)

- Move away from PeriodicLog for QueryMetricLog since in
doesn't make sense anymore.