Fix undefined behavior if all connection tries fail
For some reason, getManyCheckedForInsert was calling getManyImpl with
skip_unavailable_endpoints=nullptr, which resulted in getManyImpl using
the `skip_unavailable_shards` settings. Since this is true by default,
min_entries was set to 0.

Having min_entries set to 0 while using PoolMode::GET_ONE was strange,
to say the least. There was an edge case where if all connection
attempts failed and min_entries was 0, it was returning an empty vector.
That was not considered to be possible by the caller and it was getting
the front entry of an empty vector, causing undefined behavior.

Conclusion: set `skip_unavailable_endpoints=false` explicitly in
getManyCheckedForInsert so that min_entries=1. In case all connections
fail, an exception will be thrown. Also, add some defensive programming
to ensure we never ever try to get the front element of an empty vector.