Fix race condition in system.query_metric_log
If a scheduled task was pending to lock the query_mutex,
it could do so between the lock.unlock() and lock.lock()
of finishQuery which is used to break the exec_mutex deadlock.

In that case, a task scheduled running after the task
finished could attempt to do a new collection, which
is wrong. Let's avoid that altogether, and also add
some defensive programming to ensure this does not happen
in the future.