Fix possible overestimate memory tracking
Before this patch the untracked_memory was dumped before reseting the
current_thread which could lead to leaking some of untracked_memory in
`ThreadStatus::detachFromGroup` [1].

In particular I found that the local_data (which is 136 bytes) can be
"leaked" that way.

  [1]: https://github.com/ClickHouse/ClickHouse/blob/86f07b0cbc5704aa16e7427b0e092f3a655562dc/src/Interpreters/ThreadStatusExt.cpp#L292-L306

This pops up on one of clusters where I noticed that the difference
between MemoryTracking and MemoryResident keeps growing over the time.
the difference grows ~200KiB per second - 16.5GiB per day.

That cluster has tons MVs triggered from Buffer/Distributed (i.e. now
query context), and this is indeed is the reason for leaking local_data,
since in this case there is no thread group in the query, and it is
created explicitly in `InterpreterInsertQuery::buildChain` [1].

  [1]: https://github.com/ClickHouse/ClickHouse/blob/67a57989431419d491f4184085c909008f3d2f57/src/Interpreters/InterpreterInsertQuery.cpp#L321

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>