Disable header detection for Kafka/NATS/RabbitMQ/FileLog to improve performance
For streaming (when format is reseted after each row), any action that
is done by the format is crucial for performance, and header detection
is one of them.

Plus it does not make sense to even try it, since usually you have 1-1
mapping, i.e. 1 message -> 1 row for such formats, and I doubt that
someone depends on it.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>