ci: fix possible skipping of some integration tests
I found (in #72959, test test_logs_with_disks, see [1]) that sometimes
not all tests presented in all "Integration tests (asan, old analyzer)"
job, while all tests presented in the runner_get_all_tests.log file, not
all of them are executed.

The only place that I found is that group_test_by_file() returns dict,
in which order of keys (and hence I guess .values()) is not
deterministic (python uses some kind of seed for hash, new for each
process), and this can lead to the issue that some tests won't be
executed on any workers, since they are divided into 6 parts, and then,
each worker execute the same code, but in one worker the test can be in
"1" group, while on the other worker the same test can be on "2" group.

  [1]: https://play.clickhouse.com/play?user=play#U0VMRUNUIGNoZWNrX25hbWUsIHRlc3RfbmFtZSwgcmVwb3J0X3VybApGUk9NIGNoZWNrcwpXSEVSRSBjaGVja19uYW1lIExJS0UgJ0ludGVncmF0aW9uJScKICAgIGFuZCB0ZXN0X25hbWUgbGlrZSAnJXRlc3Rfa2VlcGVyX2Rpc2tzL3Rlc3QucHk6OnRlc3RfbG9nc193aXRoX2Rpc2tzJScKICAgIEFORCBjaGVja19zdGFydF90aW1lID49IG5vdygpIC0gSU5URVJWQUwgMjQgREFZCiAgICBhbmQgcHVsbF9yZXF1ZXN0X251bWJlciA9IDcyOTU5Ck9SREVSIEJZIGNoZWNrX25hbWUsIHRlc3RfbmFtZSwgY2hlY2tfc3RhcnRfdGltZQ==

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>