Fix applying async_insert from server (via apply_settings_from_server)
The problem is that async_insert is checked during parsing, to know what
to send to the server, even before server settings available. And this
leads to "Unknown packet 11 from server" errors.

So, first approach will be to simply ignore async_insert for
apply_settings_from_server, but this setting is interesting to apply
from the server (since this may have performance impact).

Another is to detect what part of the query we should after server
settings had been received, this patch follows this approach.