Fix Digest does not match (from Replicated database) on CI
The probelm was the 03001_matview_columns_after_modify_query test, and
only in stress tests, since there we don't set
database_replicated_always_detach_permanently=1, and so DETACH that is
done before playing with metadata on disk fails, and after this this
LOGICAL_ERROR will happen:

    $ ../tests/clickhouse-test --client-option distributed_ddl_output_mode=none --replicated-database 03001_matview_columns_after_modify_query
    ...
    Code: 49. DB::Exception: Received from localhost:9000. DB::Exception: Digest does not match. (LOGICAL_ERROR)
    (query: ALTER TABLE mv MODIFY QUERY SELECT Timestamp as timestamp, c1 || c2 as c12 FROM src)

So, let's simply disable it.

Note it does not make a lot of sense to fix this DETACH error (with i.e.
PERMANTENTLY) since we need to modify the data in ZooKeeper additionally
for Replicated, plus it still can be racy if other tests will have "mv"
table.