Introspection of .sh tests (via BASH_XTRACEFD)
Sometimes it is hard to understand why the test failed, since all .sh
tests written with a different style, and likely does not have debugging
capabilities. We have them for .sql and .expect (via exp_internal), but
not for .sh.

So the idea is to catch all the tracing (set -x) in a separate file with
BASH_XTRACEFD, that way we will get all the command invocations from
tests and match it with stderr/stdout and server logs.

Note, that I also tried to capture the stderr in the same file, but
failed (see comments in the code), but anyway this helps on it's own,
I've tried such approach before and it helps with debugging a lot.

Example:

    $ ../tests/clickhouse-test 01528_play
    ...
    01528_play:                                                             [ FAIL ] 0.24 sec.
    Reason: having stderror:
    /src/ch/clickhouse/tests/queries/0_stateless/01528_play.sh: line 7: asdf: command not found
    /src/ch/clickhouse/tests/queries/0_stateless/01528_play.sh: line 8: fail: command not found

    stdout:
    clickhouse.com

    /src/ch/clickhouse/tests/queries/0_stateless/test_6cscsqoy/01528_play.sh.debuglog:
    + [2025-04-11 10:21:15] [:7] asdf
    + [2025-04-11 10:21:15] [:8] fail
    + [2025-04-11 10:21:15] [:9] curl -q -s --max-time 120 -sS http://localhost:8123/play
    + [2025-04-11 10:21:15] [:9] grep -o -F clickhouse.com

Fixes: https://github.com/ClickHouse/ClickHouse/issues/78627
v2: remove FUNCNAME fails in case of -u