Fix possible dtor of nullptr in GraphiteRollupSortedAlgorithm dtor
The `MemorySanitizer` reports use-of-uninitialized-value of the
aggregation pattern function in GraphiteRollupSortedAlgorithm dtor. This
might happen in case `startNextGroup` called with a valid aggregation
pattern function first (then `aggregate_state_created` was set to true),
later without invoking any function in between (e.g. `insertRow`), the
`startNextGroup` function is called again with a nullptr valued
aggregation pattern function. The `aggregation_state_created` was never
set to false and was never reset to false in `startNextGroup` either.
So, we might have a nullptr valued aggregation pattern function and
`aggregation_state_created` was set to true. This might lead to invoke
the `destroy` method on the nullptr value function. This is the rare
case and probably it's false-positive report from `MemorySanitizer`.

In any case, we can replace the logic of `aggregation_state_created`
bool variable by `std::optional<Graphite::RollupRule>`.