Adjust compatibility values for max_merge_delayed_streams_for_parallel_write
The problem is that the previous value was lead to excessive memory
usage (#77922), and now if you wll use compatibility=25.3 you will have
the same.

So let's simply ignore that fact that it was 1000 at some point in time.

Note, that query-level setting
(`max_insert_delayed_streams_for_parallel_write`) was never adjusted for
compatibility, since it has default of `0` (which means auto).