Support packed storage for full text index
Previously, `packed` storage was not supported for the full-text index,
because the segment id was updated on-fly by reading and writing
(`.gin_sid`) file on disk. In case of packed storage, reading a value
from the uncommited file is not supported and this led to an issue.

To support `packed` and `SMT` storages, the segment id is now stored and
updated internally in `GinIndexStore` and saved to disk only when the
GIN index is finalized.