Extract text index search functions into its own library
To reduce the memory pressure on fuzzers, a new library
`clickhouse_functions_text` is created. Fuzzers depend on merge tree
index files and `MergeTreeIndexGin` depends on the `searchAny` and
`searchAll` functions.

Instead of including all functions (`clickhouse_functions`) in the
dependency, we can extract text index functions into its own library and
include only it in the fuzzers dependencies.