Disable filesystem cache for files that can be put into iceberg cache
After proper context was propagated for everything it started to account
more in CachedReadBufferReadFromCacheBytes/CachedReadBufferCacheWriteBytes.

This is due to first query caches metadata as well in the filesystem
cache, while subsequent does not read them.

But in general it does not make sense to use filesystem cache if there
is more precise cache, so let's simply disable it for various iceberg
metadata files.

This should fix test_storage_iceberg/test.py::test_filesystem_cache[s3]
test [1].

  [1]: https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=82146&sha=fca91d383101136506d511bad7621df8caaffb61&name_0=PR&name_1=Integration%20tests%20%28release%2C%203%2F4%29