Fix ordering issue of the FST state bitmap encoding
The 256-bit bitmap stores the outgoing labels of a state ordered, but
outgoing states are saved into disk in the order they appear in the hash
table. Therefore, a label would point to a wrong next state while
reading from disk.

To fix that, labels should be sorted and states should be saved into
disk by the same ordering.