Increase flush interval
It happened once that before the last check we polled only a single message instead of two. I am not sure why it happened, in the clickhouse logs it seems like librdkafka only returned a single message. I increased the flush interval to give CH more time to poll all the messages.