Text index: remove `CONTAINS_ALL` flag from postings list
The `CONTAINS_ALL` flag was used to save space in postings list.
When set, it indicates the term exists in the range, but exact row ids
were unknown.

Since we need to know the exact rows to benefit from text index in the
following optimizations, the flag is removed.
The postings list know contains all row ids where term exists.