Text index: disable splitting index data into multiple segments by default
Previously, the text index data would be separated into multiple
segments (each segment size by default was 256 MiB). This might reduce
the memory consumption while building the text index, however this might
increase the space requirement on the disk and increase the query
response time.

With this PR, by default we disable the that feature and make the
segment size unlimited, but still allow users to adjust if it is
necessary on the environment with the limited memory.