ci: improve gathering sanitizers reports by clickhouse-test launcher
Previously we were waiting only 10 seconds, this may not be enough,
for instance here [1], there is llvm-symbolizer process, which is a
signal that the is some issue, but no report.

  [1]: https://s3.amazonaws.com/clickhouse-test-reports/REFs/master/16a8f522f6545a8fcfb313e6bca39853caf54fc9//stateless_tests_amd_tsan_s3_storage_3_3/job.log

And note, that it should not be a problem of truncated output, since
logs from sanitizers appended to stderr, which can be truncated only via
trim_for_logs() which adds some introspection into message.

So apparently here it just fails [2], and sanitizer did not find
anything (which looks odd though).

  [2]: https://s3.amazonaws.com/clickhouse-test-reports/REFs/master/13cc4362678f9f53192a46dce0083e98a277d100//stateless_tests_amd_tsan_parallel_1_2/job.log