Fix memory tracking drift from background schedule pool and executor
It has been reported (and I also noticed it few months ago) that it is
possible to see negative MemoryTracking metric, while this should never
be possible after bunch of fixes. It is not big, ~100MiB per 1 hour, but
it still annoying.

But, huge thread pools, that reserve all threads at start may lead to
this, since each of them will have a drift up to 4MiB (it is not depends
on the `max_untracked_memory`, since we do not apply it for all
threads at start).

And indeed I've tested this and with
background_schedule_pool_size=10000, and I see that MemoryTracking can
goes down, while after this patch it is almost constant.

**And I hope this is the final bits of memory tracking drift!**

Fixes: #82036