fix flaky 03552_filesystem_read_prefetches_log
Current version may fail in case of enabled parallel replicas, when
the actual reading is performed not by initiator local pipeline.

The fix is to check prefetches for all initiated queries (also note
that current_database will be 'default' on replicas so previously
only a single query was returned as query_id_).

Also disabling merges for test table just in case, to keep 2 parts.