Fix out-of-order ActionDAG JIT compilation.
When JIT compiling an "unordered" dag like:

```
0 : COLUMN () Const(String) String 'the'_String
1 : COLUMN () Const(String) String 'can'_String
2 : FUNCTION (6, 7) (no column) UInt8 or(hasToken(__table1.text, 'the'_String), hasToken(__table1.text, 'can'_String)) [or]
3 : COLUMN () Const(String) String 'idx_1'_String
4 : INPUT () (no column) UInt64 _part_index
5 : INPUT () (no column) UInt64 _part_offset
6 : FUNCTION (3, 0, 4, 5) (no column) UInt8 hasToken(__table1.text, 'the'_String) [_hasToken_index]
7 : FUNCTION (3, 1, 4, 5) (no column) UInt8 hasToken(__table1.text, 'can'_String) [_hasToken_index]
Output nodes: 2
```

Clickhouse was hanging in an infinite loop.

This happened because of this entry

2 : FUNCTION (6, 7)

Meaning that node 2 has 6 and 7 as children. As 6 and 7 are bigger
than 2, both loops in the modified code were repeating forever.

This issue haven't been exposed before probably because the nodes in
the `ActionsDag` are always generated in order or dependency from
lefty to right. But this is actually not a real existing constrain.

* src/Interpreters/ExpressionJIT.cpp (ActionsDAG::compileFunctions) : Use apparently the correct children list.