always schedule direct io to threadpool for pread_threadpool
preadv2 with RWF_NOWAIT may perform reads from device and
return data (not EAGAIN) when called on fd with O_DIRECT
flag. This leads to data reading being executed in query
thread, instead of threadpool which is intended for
pread_threadpool.

This patch skips preadv2 / pagecache checking logic when
direct io is requested, so that it always executed in
threadpool.

Co-authored-by: Azat Khuzhin <a3at.mail@gmail.com>