Capture server stacktrace in integration tests if it hanged during shutdown
This should help with investigating why in test_global_overcommit we
cannot DROP USER at the end of the test

AFAICS we do have all logs (no drops of async logs), and in logs I don't
even see TCPHandler for this queries, so maybe there is one more
deadlock?

We can also reset server limit before running DROP, but this may just
hide the problem (even if it will help).