Fix #87887
When the input column is low cardinality, it seems like hasToken and similar functions are reported also as low cardinality result type in
the dag.

This is strictly true, but for direct read the column will contain only 0 and 1. So it is simpler and easier just set it as UInt8
explicitly in all the cases.

Maybe one day we will have some boolean support in columns (storing internally some bitmap) and we can use that specialization.