Avoid using lz4/zstd from rust crates
There is a problem with linking order, that lead to the linker prefer
symbols from crates (at least lz4/zstd are affected).

I spend some time to write "smart" cmake code that tries to put rust
dependencies at the end, but it does not work, since it cannot respect
dependencies of dependencies, and the dependecies are quiet complex
(clickhouse_common_io vs dbms libs).

I've also tried to replace some `target_link_libraries()` with
`target_include_directories(PRIVATE $<TARGET_PROPERTY:LIB,INTERFACE_INCLUDE_DIRECTORIES>)`,
but this does not help either.

So this is a temporary workaround for a problem, at least it works [1].

  [1]: https://pastila.nl/?0038ca04/550fc6bd2b159b686d00e750dd82c045#+DGmFDWDhbHDfrdMhLfZqQ==