tests: faster 00159_parallel_formatting tests
Switching to system.numbers_mt instead of `datasets.hits_v1` which
is currently used for TSAN builds, and replacing the query with
slightly simpler one (with less records to sort and less fields to
sort by).

Additionaly is looks that if HTTP checks have 1kk recordset for the
test, the other formats could use the same amount of rows.