use `CacheBase` instead of `LRUResourceCache`
Since an instance of `MergeTreeIndexGranuleText` is initialized in every
execution, the cache cannot be part of it. Instead, export to its own
class and provide a singleton API.
However, this makes the dictionary block cache a global instance, so it
needs a unique way to identify block in a granule. To do that, we pass
the part full path, index name and index mark to the granule.