Remove duplicated filesystem caches from asynchronous metrics, SYSTEM queries
It is possible to have filesystem cache i.e. "a" and "b", that points to
the same path.

Previously those was accounted multiple times in FilesystemCacheBytes
from system.asynchronous_metrics, which is wrong.

Also SYSTEM queries executed actions on the same caches multiple times
in this case, which is redundant.

Plus system.filesystem_caches may report different set of files for
various caches even though they share the same path, fix this as well.